prevent duplicate replacement

+++ src/core/replaceSourceWithTargetSource/index.ts
@@ -1,4 +1,4 @@
-import { REPLACEMENT_SOURCE_KEY } from 'debugConstants'
+import { REPLACED_KEY, REPLACEMENT_SOURCE_KEY } from 'debugConstants'
 import getTargetSource from 'utils/getTargetSource'
 import loadTargetSource from 'utils/loadTargetSource'
 import writeSourceToTargetPage from 'utils/writeSourceToTargetPage'
@@ -6,7 +6,7 @@ import writeSourceToTargetPage from 'utils/writeSourceToTargetPage'
 async function replaceSourceWithTargetSource() {
     const targetSource = getTargetSource()
 
-    if (targetSource) {
+    if (targetSource && !sessionStorage.getItem(REPLACED_KEY)) {
         console.log('Start replacing with ', targetSource)
 
         sessionStorage.setItem(REPLACEMENT_SOURCE_KEY, targetSource)
@@ -20,6 +20,8 @@ async function replaceSourceWithTargetSource() {
         sessionStorage.setItem(targetSource, source)
 
         writeSourceToTargetPage(source)
+
+        sessionStorage.setItem(REPLACED_KEY, 'true')
     }
 }
